Are people under economic stress more or less likely to vote, and why?
When unemployment is high, challengers have an
incentive to blame the incumbent, thus eliciting anger among the unemployed. Psychologists have shown
anger to be an approach or mobilizing emotion. When joblessness is low, campaigns tend to ignore it. The
jobless thus remain in states of depression and self-blame, which are demobilizing emotions.